requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
ultralytics==8.1.2
torch==2.1.2
torchvision==0.16.2 